增加開發者與其使用量說明
