+++ src/main/webapp/WEB-INF/views/developer/tutorial.jsp
@@ -1,0 +1,36 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"%>
+
+<%-- title row --%>
+<div class="row">
+    <div class="col-xs-12">
+        <h2 class="page-header">
+            <i class="fa fa-user-md"></i>開發者使用說明<small class="pull-right">修改日期：2015-06-10</small>
+        </h2>
+    </div>
+    <%-- /.col --%>
+</div>
+<%-- info row --%>
+<div class="row invoice-info">
+    <div class="col-sm-12 invoice-col">
+
+        <div id="preview" preview="">
+            <h1 id="oauth-for-">關於開發者使用</h1>
+            <p>
+                                    本功能讓開發者可以進行開發者相關操作，以便使用<strong>OAuth NCU</strong>和<strong>NCU API</strong>來開發各項應用服務。
+            </p>
+            <h5>使用限制</h5>
+            <p>每個NCU API皆有每月使用量限制，超過使用量則列入黑名單，若需要解除或有特殊需求者請來信。</p>
+            
+            <h5>一般API</h5>
+            <p>每月使用量：25000次 Requests</p>
+            
+            <h5>Personnel-Service</h5>
+            <ul>
+                <li>一分鐘試誤次數不得超過5次。</li>
+                <li>每月使用量：25000次 Requests。</li>
+            </ul>
+        </div>
+    </div>
+    <%-- /.col --%>
+</div>
+<%-- /.row --%>
